update api second deal voucher

+++ Modules/Deals/Http/Controllers/ApiDeals.php
@@ -35,6 +35,7 @@ use Modules\Deals\Entities\DealsContent;
 use Modules\Deals\Entities\DealsContentDetail;
 use Modules\Deals\Entities\DealsShipmentMethod;
 use Modules\Deals\Entities\DealsDiscountDeliveryRule;
+use Modules\Deals\Entities\SecondDealsTotal;
 
 use DB;
 
@@ -284,6 +285,8 @@ class ApiDeals extends Controller
 
         if (isset($post['is_all_days'])) {
         	$data['is_all_days'] = $post['is_all_days'];
+        }else{
+            $data['is_all_days'] = 1;
         }
 
         if (isset($post['selected_day'])) {
@@ -297,8 +300,12 @@ class ApiDeals extends Controller
     function create($data)
     {
         $data = $this->checkInputan($data);
-        $days = $data['selected_day'];
-        unset($data['selected_day']);
+
+        $days = [];
+        if (isset($data['selected_day'])) {
+            $days = $data['selected_day'];
+            unset($data['selected_day']);
+        }
 
         $data['created_by'] = auth()->user()->id;
         // error
@@ -348,7 +355,7 @@ class ApiDeals extends Controller
     }
 
     public function selectedDays($deals, $selected_days){
-    		
+        
         $table = new DealsDay;
     	$id_deals = $deals->id_deals;
 
@@ -362,10 +369,10 @@ class ApiDeals extends Controller
                 'day' 	=> $value
             ]);
         }
-
+        
         if (!empty($data_days)) {
             $save = $table::insert($data_days);
-            return $save;
+            return true;
         } else {
             return false;
         }
@@ -434,7 +441,7 @@ class ApiDeals extends Controller
             $deals->addSelect('id_brand');
             $deals->with('brand');
         }else{
-            if($request->json('deals_type') != 'WelcomeVoucher' && !$request->json('web')){
+            if($request->json('deals_type') != 'WelcomeVoucher' && $request->json('deals_type') != 'SecondDeals' && !$request->json('web')){
                 $deals->where('deals_end', '>', date('Y-m-d H:i:s'));
             }
         }
@@ -465,6 +472,7 @@ class ApiDeals extends Controller
                 $deals->addSelect('deals_promo_id','deals_publish_start','deals_publish_end','created_at');
             }
         }
+        
         if ($request->json('rule')){
              $this->filterList($deals,$request->json('rule'),$request->json('operator')??'and');
         }
@@ -494,7 +502,7 @@ class ApiDeals extends Controller
             }
         }
 
-		if ($request->json('deals_type_array')) {
+        if ($request->json('deals_type_array')) {
             // get > 1 deals types
             $deals->whereIn('deals_type', $request->json('deals_type_array'));
         }        
@@ -1123,11 +1131,17 @@ class ApiDeals extends Controller
     {
         $data = $this->checkInputan($data);
 
+        $days = [];
+        if (isset($data['selected_day'])) {
+            $days = $data['selected_day'];
+            unset($data['selected_day']);
+        }
+
         $deals = Deal::find($id);
         $data['step_complete'] = 0;
         $data['last_updated_by'] = auth()->user()->id;
 
-        if ($deals['product_type'] != $data['product_type'] || $data['is_online'] == 0) {
+        if ((isset($data['product_type'])) && ($deals['product_type'] != $data['product_type'] || $data['is_online'] == 0)) {
         	app($this->promo_campaign)->deleteAllProductRule('deals', $id);
         }
 
@@ -1165,7 +1179,22 @@ class ApiDeals extends Controller
             unset($data['id_outlet']);
         }
 
+        if($deals['is_all_days'] != $data['is_all_days'] || $data['is_all_days'] == 1){
+            DealsDay::where('id_deals', $deals['id_deals'])->delete();
+        }
+
         $save = Deal::where('id_deals', $id)->updateWithUserstamps($data);
+
+        if($data['is_all_days'] == 0 && isset($days)){
+            $saveDays = $this->selectedDays($deals, $days);
+            if (!$saveDays) {
+                DB::rollBack();
+                $result = [
+                    'status'  => 'fail',
+                    'message'  => ['Failed to create deals days']
+                ];
+            }
+        }
 
         return $save;
     }
@@ -1203,6 +1232,9 @@ class ApiDeals extends Controller
 	                    break;
 	                case 'welcomevoucher':
 	                    $dt = 'Welcome Voucher';
+	                    break;
+	                case 'seconddeals':
+	                    $dt = 'Second Deals';
 	                    break;
 	            }
 	            $deals = Deal::where('id_deals',$request->json('id_deals'))->first()->toArray();
@@ -1511,7 +1543,7 @@ class ApiDeals extends Controller
     	}
 
         if ( ($post['step'] == 1 || $post['step'] == 'all') && ($deals_type != 'Promotion') ){
-			$deals = $deals->with(['outlets']);
+			$deals = $deals->with(['outlets','deals_days']);
         }
 
         if ($post['step'] == 2 || $post['step'] == 'all') {
@@ -1522,8 +1554,11 @@ class ApiDeals extends Controller
                 $table.'_tier_discount_rules',
                 $table.'_buyxgety_product_requirement',
                 $table.'_buyxgety_rules.product',
+                $table.'_productcategory_category_requirements',
+                $table.'_productcategory_rules.product_category',
                 $table.'_shipment_method',
-                $table.'_discount_delivery_rules'
+                $table.'_discount_delivery_rules',
+                'deals_days'
             ]);
         }
 
@@ -1654,7 +1689,7 @@ class ApiDeals extends Controller
     	$deals = $deals->toArray();
     	if ( $deals['is_online'] == 1)
     	{
-	    	if ( empty($deals['deals_product_discount_rules']) && empty($deals['deals_tier_discount_rules']) && empty($deals['deals_buyxgety_rules']) && empty($deals['deals_discount_delivery_rules']))
+	    	if ( empty($deals['deals_product_discount_rules']) && empty($deals['deals_tier_discount_rules']) && empty($deals['deals_buyxgety_rules']) && empty($deals['deals_discount_delivery_rules']) && empty($deals['deals_productcategory_rules']))
 	    	{
 	    		$step = 2;
 	    		$errors = 'Deals not complete';
@@ -2611,5 +2646,98 @@ class ApiDeals extends Controller
             ];
         }
         return response()->json($result);
+    }
+
+    function listDealsSecondDeals(Request $request){
+        $configUseBrand = Configs::where('config_name', 'use brand')->first();
+
+        if($configUseBrand['is_active']){
+            $getDeals = Deal::join('brands', 'brands.id_brand', 'deals.id_brand')
+                ->where('deals_type','SecondDeals')
+                ->select('deals.*','brands.name_brand')
+                ->get()->toArray();
+        }else{
+            $getDeals = Deal::where('deals_type','SecondDeals')
+                ->select('deals.*')
+                ->get()->toArray();
+        }
+
+
+        $result = [
+            'status' => 'success',
+            'result' => $getDeals
+        ];
+        return response()->json($result);
+    }
+
+    function secondDealsSetting(Request $request){
+        $setting = Setting::where('key', 'second_deals_setting')->first();
+        $configUseBrand = Configs::where('config_name', 'use brand')->first();
+
+        if($configUseBrand['is_active']){
+            $getDeals = SecondDealsTotal::join('deals', 'deals.id_deals', 'second_deals_totals.id_deals')
+                ->join('brands', 'brands.id_brand', 'deals.id_brand')
+                ->select('deals.*','second_deals_totals.deals_total','brands.name_brand')
+                ->get()->toArray();
+        }else{
+            $getDeals = SecondDealsTotal::join('deals', 'deals.id_deals', 'second_deals_totals.id_deals')
+                ->select('deals.*','second_deals_totals.deals_total')
+                ->get()->toArray();
+        }
+
+
+        $result = [
+            'status' => 'success',
+            'data' => [
+                'setting' => $setting,
+                'deals' => $getDeals
+            ]
+        ];
+        return response()->json($result);
+    }
+
+    function secondDealsSettingUpdate(Request $request){
+        $post = $request->json()->all();
+
+        $deleteDealsTotal = DB::table('second_deals_totals')->delete();//Delete all data from tabel deals total
+
+        //insert data
+        $arrInsert = [];
+        $list_id = $post['list_deals_id'];
+        $list_deals_total = $post['list_deals_total'];
+        $count = count($list_id);
+
+        for($i=0;$i<$count;$i++){
+            $data = [
+                'id_deals' => $list_id[$i],
+                'deals_total' => $list_deals_total[$i],
+                'created_at' => date('Y-m-d H:i:s'),
+                'updated_at' => date('Y-m-d H:i:s'),
+                'created_by' => Auth::id(),
+            	'updated_by' => Auth::id()
+            ];
+            array_push($arrInsert,$data);
+        }
+
+        $insert = SecondDealsTotal::insert($arrInsert);
+        if($insert){
+            $result = [
+                'status' => 'success'
+            ];
+        }else{
+            $result = [
+                'status' => 'fail'
+            ];
+        }
+
+        return response()->json($result);
+    }
+
+    function secondDealsSettingUpdateStatus(Request $request){
+        $post = $request->json()->all();
+        $status = $post['status'];
+        $updateStatus = Setting::where('key', 'second_deals_setting')->updateWithUserstamps(['value' => $status]);
+
+        return response()->json(MyHelper::checkUpdate($updateStatus));
     }
 }
